Refactoring: splitting big files into smaller abstractions.
