+++ Enigmatry.Entry.CodeGeneration.Configuration/Form/Controls/ControlTypes.cs
@@ -13,18 +13,3 @@ public static class ControlTypes
     public const string TextArea = "textarea";
     public const string Button = "button";
 }
-
-[Obsolete("Use ControlTypes instead.")]
-public static class FormlyTypes
-{
-    public const string Input = ControlTypes.Input;
-    public const string CheckBox = ControlTypes.CheckBox;
-    public const string DatePicker = ControlTypes.DatePicker;
-    public const string DateTimePicker = ControlTypes.DateTimePicker;
-    public const string Radio = ControlTypes.Radio;
-    public const string Select = ControlTypes.Select;
-    public const string Autocomplete = ControlTypes.Autocomplete;
-    public const string MultiCheckBox = ControlTypes.MultiCheckBox;
-    public const string TextArea = ControlTypes.TextArea;
-    public const string Button = ControlTypes.Button;
-}
