+++ Enigmatry.Entry.CodeGeneration.Configuration/Form/Controls/FormlyTypes.cs
@@ -1,0 +1,16 @@
+namespace Enigmatry.Entry.CodeGeneration.Configuration.Form.Controls;
+
+[Obsolete("Use ControlTypes instead.")]
+public static class FormlyTypes
+{
+    public const string Input = ControlTypes.Input;
+    public const string CheckBox = ControlTypes.CheckBox;
+    public const string DatePicker = ControlTypes.DatePicker;
+    public const string DateTimePicker = ControlTypes.DateTimePicker;
+    public const string Radio = ControlTypes.Radio;
+    public const string Select = ControlTypes.Select;
+    public const string Autocomplete = ControlTypes.Autocomplete;
+    public const string MultiCheckBox = ControlTypes.MultiCheckBox;
+    public const string TextArea = ControlTypes.TextArea;
+    public const string Button = ControlTypes.Button;
+}
