+++ Enigmatry.Entry.CodeGeneration.Validation/ValidationRules/IFluentValidationValidationRule.cs
@@ -1,0 +1,8 @@
+using System.Linq.Expressions;
+
+namespace Enigmatry.Entry.CodeGeneration.Validation.ValidationRules;
+
+public interface IFluentValidationValidationRule : IBaseValidationRule
+{
+    LambdaExpression Expression { get; }
+}
