+++ Enigmatry.Entry.CodeGeneration.Validation/ValidationRules/IFormlyValidationRule.cs
@@ -1,0 +1,10 @@
+namespace Enigmatry.Entry.CodeGeneration.Validation.ValidationRules;
+
+public interface IFormlyValidationRule : IBaseValidationRule
+{
+    string ValidationMessage { get; }
+    string RuleName { get; }
+    string[] TemplateOptions { get; }
+
+    void SetMessageTranslationId(string messageTranslationId);
+}
