+++ Enigmatry.Entry.CodeGeneration/Templates/HtmlHelperExtensions/Angular/AngularSignalsSelectHtmlHelperExtensions.cs
@@ -1,0 +1,75 @@
+using Enigmatry.Entry.CodeGeneration.Configuration.Form;
+using Enigmatry.Entry.CodeGeneration.Configuration.Form.Controls;
+using Enigmatry.Entry.CodeGeneration.Templates.HtmlHelperExtensions.TypeScript;
+using Microsoft.AspNetCore.Html;
+using Microsoft.AspNetCore.Mvc.Rendering;
+
+namespace Enigmatry.Entry.CodeGeneration.Templates.HtmlHelperExtensions.Angular;
+
+public static class AngularSignalsSelectHtmlHelperExtensions
+{
+    public static IHtmlContent SelectInputDeclarations(this IHtmlHelper htmlHelper, SelectControlBase select, bool enableI18N)
+    {
+        var lines = new List<string>();
+
+        if (select.Options.HasDynamicValues)
+        {
+            lines.Add($"    protected readonly {select.PropertyName}Callback = input<Observable<unknown[]> | null>(null);");
+            lines.Add($"    private readonly {select.PropertyName}RawOptions = toSignal(toObservable(this.{select.PropertyName}Callback).pipe(switchMap(callback => callback ?? of([]))), {{ initialValue: [] }});");
+        }
+        else
+        {
+            var options = htmlHelper.JsArray(select.Options.FixedOptions,
+                option => $"{{ {option.GetValueExpression()}, displayName: {htmlHelper.Localize(option.DisplayName, enableI18N)} }}");
+            lines.Add($"    private readonly {select.PropertyName}RawOptions = signal({options});");
+        }
+
+        if (select.Options.HasCustomValueAndDisplayKeys)
+        {
+            lines.Add($"    protected readonly {select.PropertyName}OptionsConfiguration = input<SelectConfiguration>({{ valueProperty: '{select.Options.OptionValueKey}', labelProperty: '{select.Options.OptionDisplayKey}', sortProperty: '{select.Options.OptionSortKey}' }});");
+        }
+        else
+        {
+            lines.Add($"    protected readonly {select.PropertyName}OptionsConfiguration = input<SelectConfiguration>({select.Options.DefaultOptionsAsString});");
+        }
+
+        lines.Add($"    protected readonly {select.PropertyName}Options = computed(() => {{");
+        lines.Add($"        const configuration = this.{select.PropertyName}OptionsConfiguration();");
+        lines.Add($"        const options = sortOptions(this.{select.PropertyName}RawOptions() as Record<string, unknown>[], configuration.valueProperty ?? 'value', configuration.sortProperty ?? '', this.localeId) as Record<string, unknown>[];");
+        lines.Add($"        return options.map(option => ({{ value: option[configuration.valueProperty ?? 'value'], displayName: option[configuration.labelProperty ?? 'displayName'] }}));");
+        lines.Add($"    }});");
+
+        if (select is AutocompleteFormControl)
+        {
+            var propertyNameCapitalized = AngularSignalsFormModelExtensions.Capitalize(select.PropertyName);
+            lines.Add($"    protected readonly display{propertyNameCapitalized} = (value: unknown): string =>");
+            lines.Add($"        (this.{select.PropertyName}Options().find(option => option.value === value)?.displayName as string) ?? '';");
+        }
+
+        return htmlHelper.Raw(String.Join("\r\n", lines) + "\r\n");
+    }
+
+    public static IHtmlContent AllSelectInputDeclarations(this IHtmlHelper htmlHelper, FormComponentModel model, bool enableI18N) =>
+        htmlHelper.Raw(String.Concat(model.FormControlsOfType<SelectControlBase>()
+            .Select(select => htmlHelper.SelectInputDeclarations(select, enableI18N).ToString())));
+
+    public static IHtmlContent AllAutocompleteFilterDeclarations(this IHtmlHelper htmlHelper, FormComponentModel model)
+    {
+        var autocompleteControls = model.FormControlsOfType<AutocompleteFormControl>().ToList();
+        if (autocompleteControls.Count == 0)
+        {
+            return htmlHelper.Raw("");
+        }
+
+        var declarations = autocompleteControls.Select(autocomplete =>
+            $"    private readonly {autocomplete.PropertyName}FilterValue = toSignal(this.form.controls.{autocomplete.PropertyName}.valueChanges, {{ initialValue: null }});\r\n" +
+            $"    protected readonly {autocomplete.PropertyName}FilteredOptions = computed(() => {{\r\n" +
+            $"        const filterValue = this.{autocomplete.PropertyName}FilterValue();\r\n" +
+            $"        const filterText = typeof filterValue === 'string' ? filterValue.toLowerCase() : '';\r\n" +
+            $"        return this.{autocomplete.PropertyName}Options().filter(option =>\r\n" +
+            $"            option.value === filterValue || String(option.displayName).toLowerCase().includes(filterText));\r\n" +
+            $"    }});\r\n");
+
+        return htmlHelper.Raw("\r\n" + String.Concat(declarations));
+    }
+}
